codecs: minor refactoring
Refactor codec API so that a buffer size argument is always directly after
the buffer pointer argument.

Use uint8_t instead of unsigned char for generic data buffers, because a char
is not always 8 bits (16 bits on some embedded platforms).

Use size_t instead of int or unsigned int when the represented value is a size
in bytes.

+++ daemon/src/audio/audiortp/audio_rtp_session.cpp
@@ -173,7 +173,7 @@ void AudioRtpSession::receiveSpeakerData()
 
 void AudioRtpSession::sendMicData()
 {
-    int compSize = rtpStream_.processDataEncode();
+    size_t compSize = rtpStream_.processDataEncode();
 
     // if no data return
     if (compSize == 0)
